add changes

+++ mlkem/poly.c
@@ -308,7 +308,7 @@ void poly_tomsg(uint8_t msg[KYBER_INDCPA_MSGBYTES], const poly *a) {
   for (i = 0; i < KYBER_N / 8; i++) {
     msg[i] = 0;
     for (j = 0; j < 8; j++) {
-      t = a->coeffs[8 * i + j];
+      t = (uint32_t)a->coeffs[8 * i + j];
       // t += ((int16_t)t >> 15) & KYBER_Q;
       // t  = (((t << 1) + KYBER_Q/2)/KYBER_Q) & 1;
       t <<= 1;
